requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
pyyaml==6.0.2
pydantic==2.9.0
